vdk-impala: add Out Of Memory error handling
Address review feedback
